+++ tests/acceptance/default-values-test.js
@@ -1,98 +1,96 @@
-import { run } from '@ember/runloop';
-import $ from 'jquery';
-import {
-  describe,
-  it,
-  beforeEach,
-  afterEach
-} from 'mocha';
+import xSelectInteractor from 'dummy/tests/helpers/x-select';
+import pageInteractor from 'dummy/tests/interactors/test-page';
 import { expect } from 'chai';
-import startApp from '../helpers/start-app';
-import { select } from 'dummy/tests/helpers/x-select';
+import { visit } from '@ember/test-helpers';
+import { when } from '@bigtest/convergence';
+import { describe, it, beforeEach } from 'mocha';
+import { setupApplicationTest } from 'ember-mocha';
 
 describe('XSelect: Default Values', function() {
-  let application;
+  let page = new pageInteractor();
+  let makeSelect = new xSelectInteractor('.spec-car-make');
+  let modelSelect = new xSelectInteractor('.spec-car-model');
+  let trimSelect = new xSelectInteractor('.spec-car-trim');
+  let modelMakeSelect = new xSelectInteractor('.spec-autopopulated-make-field');
+  let quantitySelect = new xSelectInteractor('.spec-autopopulated-quantity-field');
 
-  beforeEach(function() {
-    application = startApp();
-  });
+  setupApplicationTest();
 
-  afterEach(function() {
-    run(application, 'destroy');
-  });
-
-  describe('Initializing with default values', function(){
-    beforeEach(function() {
-      visit('test-bed/default-value');
+  describe('Initializing with default values', function() {
+    beforeEach(async function() {
+      await visit('test-bed/default-value');
     });
 
-    it('initializes with defaults if no explicit value is present', function() {
-      expect($(".spec-selected-make:contains('Selected Make: Honda')")).to.exist;
+    it('initializes with defaults if no explicit value is present', async () => {
+      await when(() => expect(page.carMakeText).to.equal('Selected Make: Honda'));
     });
 
-    it('sets the selected property on the correct default option', function() {
-      expect($(".spec-car-make option:contains('Honda')")).to.be.selected;
+    it('sets the selected property on the correct default option', async () => {
+      await when(() => expect(makeSelect.options(1).isSelected).to.equal(true));
     });
 
-    it('can set a default to the first option in a dynamic list', function() {
-      expect($(".spec-selected-model:contains('Selected Model: Fit')")).to.exist;
+    it('can set a default to the first option in a dynamic list', async () => {
+      await when(() => expect(page.carModelText).to.equal('Selected Model: Fit'));
     });
 
-    it('sets the selected property on the correct default option', function() {
-      expect($(".spec-car-model option:contains('Fit')")).to.be.selected;
+    it('sets the selected property on the correct default option', async () => {
+      await when(() => expect(modelSelect.options(0).isSelected).to.equal(true));
     });
 
-    it('initializes with the correct explicit value if one is present', function() {
-      expect($(".spec-selected-make-from-model:contains('Selected Make: Ford')")).to.exist;
+    it('initializes with the correct explicit value if one is present', async () => {
+      await when(() => expect(modelMakeSelect.options(2).isSelected).to.equal(true));
     });
 
-    it('sets the selected property on the correct explicity value option', function() {
-      expect($(".spec-autopopulated-make-field option:contains('Ford')")).to.be.selected;
+    it('sets the selected property on the correct explicity value option', async () => {
+      await when(() => expect(modelMakeSelect.options(2).text).to.equal('Ford'));
     });
 
-    it('does not set the selected property on the default option', function() {
-      expect($(".spec-autopopulated-make-field option:contains('Toyota')")).not.to.be.selected;
+    it('does not set the selected property on the default option', async () => {
+      await when(() => expect(modelMakeSelect.options(0).isSelected).to.equal(false));
     });
 
-    it('sets the selected property to the explicitly set value', function() {
-      expect($(".spec-autopopulated-quantity-field option:contains('0')")).to.be.selected;
+    it('sets the selected property to the explicitly set value', async () => {
+      await when(() => {
+        expect(quantitySelect.options(5).text).to.equal('0');
+        expect(quantitySelect.options(5).isSelected).to.equal(true);
+      });
     });
 
-    it('initializes with the correct explicit value if one is present even if that value is falsy', function() {
-      expect($(".spec-selected-quantity:contains('Selected Quantity: 0')")).to.exist;
+    it('initializes with the correct explicit value if one is present even if that value is falsy', async () => {
+      await when(() => expect(page.selectedQuantityText).to.equal('Selected Quantity: 0'));
     });
 
-    it('sets the selected property on the correct default option', function() {
-      expect($(".spec-car-trim option:contains('Sport')")).to.be.selected;
+    it('sets the selected property on the correct default option', async () => {
+      await when(() => {
+        expect(trimSelect.options(0).isSelected).to.equal(true);
+        expect(trimSelect.options(0).text).to.equal('Sport');
+      });
     });
 
-    it('initializes with the correct explicit value if one is present even if that value is undefined', function() {
-      expect($(".spec-selected-trim:contains('Selected Trim: Sport')")).to.exist;
-    });
-
-    describe("Changing the value on fields with default values", function() {
-      beforeEach(function() {
-        select('.spec-car-make', 'Toyota');
+    describe('Changing the value on fields with default values', function() {
+      beforeEach(async () => {
+        await makeSelect.select('Toyota');
       });
 
-      it("updates the value", function() {
-        expect($(".spec-selected-make:contains('Selected Make: Toyota')")).to.exist;
+      it('updates the value', async () => {
+        await when(() => {
+          expect(makeSelect.options(2).isSelected).to.equal(true);
+          expect(makeSelect.options(2).text).to.equal('Toyota');
+        });
       });
 
-      it("sets the selected property on the correct updated option", function() {
-        expect($(".spec-car-make option:contains('Toyota')")).to.be.selected;
+      it('removes the selected property on the previously selected option', async () => {
+        await when(() => {
+          expect(makeSelect.options(1).isSelected).to.equal(false);
+          expect(makeSelect.options(1).text).to.equal('Honda');
+        });
       });
 
-      it("removes the selected property on the previously selected option", function() {
-        expect($(".spec-car-make option:contains('Honda')")).not.to.be.selected;
-      });
-
-      it("reevalutates the dynamic default value", function() {
-        expect($(".spec-selected-model:contains('Selected Model: Camry')")).to.exist;
-      });
-
-      it("sets the selected property on the correct default value", function() {
-        expect($(".spec-car-model option:contains('Camry')")).to.be.selected;
+      it('reevalutates the dynamic default value', async () => {
+        await when(() => {
+          expect(modelSelect.options(0).isSelected).to.equal(true);
+          expect(modelSelect.options(0).text).to.equal('Camry');
+        });
       });
     });
   });