+++ tests/dummy/app/mixins/folks.js
@@ -1,10 +1,10 @@
 import Mixin from '@ember/object/mixin';
 
-export const bastion = {name: 'Bastion', cowboy: 'nope'};
+export const bastion = { name: 'Bastion', cowboy: 'nope' };
 
-export const charles = {name: 'Charles', cowboy: 'yep'};
+export const charles = { name: 'Charles', cowboy: 'yep' };
 
-export const stanley = {name: 'Stanley', cowboy: 'maybe'};
+export const stanley = { name: 'Stanley', cowboy: 'maybe' };
 
 export default Mixin.create({
   bastion: bastion,