+++ tests/helpers/start-app.js
@@ -2,7 +2,6 @@ import Application from '../../app';
 import config from '../../config/environment';
 import { merge } from '@ember/polyfills';
 import { run } from '@ember/runloop';
-import './get-component-by-id';
 
 export default function startApp(attrs) {
   let attributes = merge({}, config.APP);