+++ tests/integration/components/x-select-actions-test.js
@@ -1,23 +1,25 @@
-import Ember from 'ember';
 import { expect } from 'chai';
+import { when } from '@bigtest/convergence';
 import { setupComponentTest } from 'ember-mocha';
 import { beforeEach, describe, it } from 'mocha';
-import { select } from 'dummy/tests/helpers/x-select';
-import hbs from 'htmlbars-inline-precompile';
 import sinon from 'sinon';
+import hbs from 'htmlbars-inline-precompile';
+import xSelectInteractor from 'dummy/tests/helpers/x-select';
 
 describe('Integration: XSelectActionsComponent', function() {
+  let xselect = new xSelectInteractor('.x-select');
+
   setupComponentTest('x-select-actions', {
     integration: true
   });
 
-  describe("x-option actions", function() {
+  describe('x-option actions', function() {
     beforeEach(function() {
       this.set('handleDisable', sinon.spy());
 
       this.render(hbs`
         {{#x-select value=value as |xs|}}
-          {{#xs.option value="Hello" on-disable=(action handleDisable) disabled=disabledProp class="test-xs-option"}}
+          {{#xs.option value="Hello" on-disable=(action handleDisable) disabled=disabledProp}}
             Hello
           {{/xs.option}}
         {{/x-select}}
@@ -25,33 +27,39 @@ describe('Integration: XSelectActionsComponent', function() {
       this.set('disabledProp', false);
     });
 
-    it('the option is not disabled', function() {
-      expect(this.$('.test-xs-option').prop('disabled')).to.equal(false);
+    it('the option is not disabled', async () => {
+      await when(() => {
+        expect(xselect.options(0).isDisabled).to.equal(false);
+        expect(xselect.options(0).text).to.equal('Hello');
+      });
     });
 
-    it("does not fire the action", function() {
+    it('does not fire the action', function() {
       expect(this.get('handleDisable')).to.not.have.been.called;
     });
 
-    describe("disabling an option", function() {
+    describe('disabling an option', function() {
       beforeEach(function() {
         this.set('disabledProp', true);
       });
 
-      it("disables the option", function() {
-        expect(this.$('.test-xs-option').prop('disabled')).to.equal(true);
+      it('disables the option', async () => {
+        await when(() => expect(xselect.options(0).isDisabled).to.equal(true));
       });
 
-      it("calls the disable action", function() {
+      it('calls the disable action', function() {
         expect(this.get('handleDisable')).to.have.been.called;
       });
 
-      it("has the correct arguments passed to the action", function() {
+      it('has the correct arguments passed to the action', function() {
         expect(this.get('handleDisable').args[0][0]).to.equal('Hello', 'First argument should be the value');
-        expect(this.get('handleDisable').args[0][1]).to.equal(true, 'Second argument should be the disabled boolean');
+        expect(this.get('handleDisable').args[0][1]).to.equal(
+          true,
+          'Second argument should be the disabled boolean'
+        );
       });
 
-      describe("setting the same value", function() {
+      describe('setting the same value', function() {
         beforeEach(function() {
           this.set('disabledProp', true);
         });
@@ -61,29 +69,6 @@ describe('Integration: XSelectActionsComponent', function() {
           // parent describe
           expect(this.get('handleDisable')).to.have.been.calledOnce;
         });
-
-      });
-
-    });
-  });
-
-  describe("x-select actions", function() {
-    describe("sending a string action", function() {
-      beforeEach(function() {
-        this.render(hbs`
-        {{#x-select value=value on-change="stringaction" as |xs|}}
-          {{#xs.option value="Hello"}}Hello{{/xs.option}}
-        {{/x-select}}
-      `);
-        /* eslint-disable */
-        Ember.warn = sinon.spy();
-
-        select(this.$(), 'Hello');
-      });
-
-      it("sends a warning in the console", function() {
-        expect(Ember.warn).to.have.been.calledOnce;
-        /* eslint-enable */
       });
     });
   });